Move integration tests out of unit tests (#80)
* exploring integration testing for binaries

https://doc.rust-lang.org/book/ch11-03-test-organization.html

* wip

* more other tests

* add a local dev crate with shared function between unit and integration tests

https://stackoverflow.com/questions/44539729/what-is-an-idiomatic-way-to-have-shared-utility-functions-for-integration-tests

* remove old code

* run code coverage against workspace

* update script + instructions for coverage tests

* rename to test_helpers

* Update integration_get.rs

* Update pyrsia_node/readme.md


* mergo conflict
