+++ pyrsia_node/src/network.rs
@@ -1,3 +1,4 @@
 pub mod behavior;
+pub mod http;
 pub mod swarm;
 pub mod transport;
