+++ tests/code_coverage.Dockerfile
@@ -5,6 +5,6 @@ RUN rustup --version; \
     cargo --version; \
     rustc --version; \
     cargo install cargo-tarpaulin;
-COPY pyrsia_node/ pyrsia
-WORKDIR pyrsia
-CMD ["cargo", "tarpaulin", "-v"]
+COPY . /home/pyrsia/
+WORKDIR /home/pyrsia
+CMD ["cargo", "tarpaulin", "--workspace"]
